+++ src/renderer/callbacks/intervaled.ts
@@ -12,10 +12,9 @@ import { getOriginIdFromName } from '../screens/OpenGov/utils';
 import { getTracks } from '@/model/Track';
 import { getMinApprovalSupport, rmChars } from '../utils/openGov';
 import type { AnyData } from '@/types/misc';
-import type { ActiveReferendaInfo } from '@/types/openGov';
+import type { ActiveReferendaInfo, OneShotReturn } from '@/types/openGov';
 import type { IntervalSubscription } from '@/controller/renderer/IntervalsController';
 import type { NotificationData } from '@/types/reporter';
-import type { OneShotReturn } from '@/model/Track';
 
 /// Debugging function.
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
